refactor : kakaoClient의 모듈 위치 수정
